final working version + test
