+++ apm-agent-plugins/apm-es-restclient-plugin/apm-es-restclient-plugin-7_x/src/main/resources/META-INF/services/co.elastic.apm.agent.sdk.ElasticApmInstrumentation
@@ -1,0 +1,1 @@
+co.elastic.apm.agent.esrestclient.v7_x.RestClientInstrumentation
